Autoconnect Newly Created Subgraph Nodes (#1289)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

A continuation of the Nodes -> Subgraph work.

This PR adds auto-reconnection functionality to newly created subgraph nodes.

Any connections in and out of the newly created subgraph node will be retained and redirected to the subgraph's task inputs and outputs. The result is that the node will seamless appear, fully connected, in the current graph!

To make this work `addTask` needed to be updated:

- Fix a bug where arguments were no properly copied over when adding a task
- `addTask` now returns the generated taskId.

Values of input nodes captured within the selection will also be moved over to the new subgraph's task arguments.

## Related Issue and Pull requests

Follows #1231

<!-- Link to any related issues using the format #<issue-number> -->

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] New feature
- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

[subgraph-autoconnection.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.com/user-attachments/thumbnails/9fd54576-c357-4197-a72a-4eb0d9599a37.mov" />](https://app.graphite.com/user-attachments/video/9fd54576-c357-4197-a72a-4eb0d9599a37.mov)

## Test Instructions

1. create a pipeline with tasks, inputs and outputs
2. select some of the nodes and create a subgraph out of them
3. the new subgraph node should automatically maintain connections to the res of the graph
4. input node values within the selection should migrate over to the subgraph node's task arguments
5. everything else should work as expected (including run submission)

<!-- Detail steps and prerequisites for testing the changes in this PR -->

## Additional Comments

Note: input nodes that are moved into the subgraph will have their value extracted and copied into the subgraph node's task arguments (i.e. the task input will be autopopulated). This means that the input node inside the subgraph has no assigned value. 

Currently all input nodes must have a value so this is technically incorrect. However, this still works because Input nodes are intended to be optional - but we have temporarily disabled the UI functionality for this. Once the nodes -> subgraph work is complete we will revisit Input Node optionality and can update the methods here accordingly. 

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->