Prompt User for Subgraph Name (#1305)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

The user will now be prompted to name their subgraph when they create it.

This is done via a new `InputDialog` component that pauses the execution and waits for the user to enter & confirm a value. It also supports validations on the entered string and additional content below the input box (which is used here to list the impacted nodes for additional context). `InputDialog`​ follows the same pattern as `ConfirmationDialog`​, with a hook to handle all of the internal works of the dialog.

The name field will validate against the input validation function, which disallows empty names, duplicates and special characters.

`NodesList`​ has been added for displaying a list of nodes in a collapsible. This is used in the dialog if the user wants to easily refer to the `node_id`​ of the nodes they are converted, perhaps for subgraph naming inspiration. The dots on the node list correspond with task colours (blue = input etc)

Additionally, to provide better separation of concerns and improved clarity, as well as reducing bloat in `FlowCanvas`​, the business logic of `onGroupNodes`​ has been moved out of `FlowCanvas`​ and into a new `callbacks`​ directory. In future we can seek to move other major operations (e.g. duplication, replacements, upgrade) out of `FlowCanvas`​ and into their own files for improved readability and testability.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

Completes the core functionality of the Nodes -> Subgraph work

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] New feature
- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

![image.png](https://app.graphite.com/user-attachments/assets/05b5fd8e-f608-45ca-8e0b-ec5daa5e48f3.png)

![image.png](https://app.graphite.com/user-attachments/assets/e445ae58-2534-46c3-abd8-b663b0720edb.png)



## Test Instructions

1. Select a group of nodes
2. Convert to subgraph
3. Subgraph naming dialog should open
    1. name field should be prepopulated with a default
    2. collapsible box should show the correct list of affected nodes
    3. name field does not allow special characters, empty names, or duplicates
    4. "Cancel" button should close the dialog and abort the procedure
    5. "Continue" button should proceed and a new subgraph should appear with the input name

<!-- Detail steps and prerequisites for testing the changes in this PR -->

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->