Rename Subgraphs (#1315)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

Allows users to rename subgraph components after they have been previously named & created. This is accessible through a new edit button next to the subgraph name.

This is done via the same `InputDialog` component and validation logic for the original naming action.

Note: When a subgraph is initially named we check the input name is a unique task id. On subsequent renaming we keep the id static and only check if the name itself is unique. Arguably we do not need to enforce unique task names (only ids), so I am open to removing this restriction.



## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] New feature
- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

![image.png](https://app.graphite.com/user-attachments/assets/b0df9430-68d0-423d-aef8-f77fa4f98bb7.png)

![image.png](https://app.graphite.com/user-attachments/assets/73e50042-5ccf-4052-9891-e4c8b6e55d45.png)

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
1. create a subgraph
2. in the task context panel click the rename button next to the task name
3. dialog should open and prompt for input
4. enter a new name and confirm
5. subgraph name should be updated on the canvas and within the pipeline spec
6. the above flow should not be available on the run page

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->

This same logic should work perfectly fine for normal tasks as well, but for now I have that disabled.