Improvements to IO Node Ghosts & Quick-Place (#1440)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
IO Ghost nodes now:
- Measure their own size and position themselves correctly (instead of working with an assumed default)
- End up in the same position when placed
- Have the connection line drawn to their handle, not the top of the node
- No longer append "Input" or "Output" to the name when placed

To enable this a custom connection line component was added. This component renders for all in-progress connections, except when a GhostNode is active. This is used in conjunction with a GhostEdge to draw the line to the ghost node's handle instead of the cursor.

Additionally, Edges:
- The default drawn edge is now similar in style to our placed edges (i.e. thicker)
- Now color a slight green whilst drawn when valid (e.g. task -> task)
- Now color a slight red whilst drawn when invalid (e.g. input -> output)
- Now color a dashed blue whilst drawn when placing a ghost Input
- Now color a dashed purple whilst drawn when placing a ghost Output

Finally, some additional code improvements were made to the Ghost Node logic. e.g. consolidating refs and removing and expensive `nodes.find` statement.

Note: due to our existing connection architecture the IO Node handles cannot have ids, but for the ghost node connections they need an id. Hence, there is a ternary to separate the two cases. In future we should revise (e.g. as per node manager stack) and all handles should have ids.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Bug fix
- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
Confirm that edges can still be drawn as expected
Confirm that ghost nodes position and move and behave as expected
Confirm that ghost nodes place in the same location as they were when dropped
Confirm that drawing a connection goes green when valid and red when invalid

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->