InlineStack Default to BlockAlign-Center (#1484)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
A majority of our `InlineStack`s are using `blockAlign="center"` or putting identical size things in a row, so we might as well make it default.


It is VERY LIKELY that this change will result in a few cases where things are unintentionally misaligned, but we can pick those up and fix them as we go.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] (Potentially) Breaking change

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->

App should look exactly the same.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->

We currently use `InlineStack` 136 times in the codebase.
Of that, 84 of them use `blockAlign="center"`.