Add RunID to Cloned Pipelines (#1560)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Cloning a pipeline from a run will now add a `cloned_from_run_id` annotation to the new pipeline.

Note: this is bugged for the "Clone Pipeline" button in the top menu bar when in a subgraph because the url params for a subgraph switch to execution id instead of run id. Not sure yet what the best solution is - in the past we've made a conscious effort to avoid API calls in the header (so we can't use the execution id to call for the run id)

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

Closes https://github.com/TangleML/tangle-ui/issues/260

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

![image.png](https://app.graphite.com/user-attachments/assets/92257e8d-caee-4408-b896-9169951fe2c6.png)

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
verify that the annotation correctly appears when a run is cloned.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->