Argument Editor Visual Update (#1558)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
A quick visual refresh of the Argument Editor to make the input fields multiple lines, like the Annotation Editor. Moved the action buttons into the input box hover for a cleaner look. Removed the description show/hide toggle - if a description is provided it will always be visible.

NOTE: I did not want to spend much time on this as my primary goal was to fix the overflowing/wrapping buttons. I went for a quick visual rehash rather than a full UX/cleanup rework. All buttons and functionality is the same, just the UI/layout has changed.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->
Closes https://github.com/TangleML/tangle-ui/issues/1333

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Bug fix
- [x] Improvement
- [x] Cleanup/Refactor

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

Before:

![image.png](https://app.graphite.com/user-attachments/assets/ab1bd071-dc39-42bc-9d9e-2accf261b60c.png)

After:

![image.png](https://app.graphite.com/user-attachments/assets/ca7b2c3b-04be-4253-9490-a761535f4a73.png)

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
Verify that the Argument Editor still works as it did previously.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->