Add UI Blocks to Context Panel (#1470)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Adds standard UI blocks for use inside the ContextPanel. Hopefully this will allow us a bit more visual consistency going forward. Each block has a header/title and some content. The header is the same for all blocks, but the content will be different depending on the block.

Implementation of the Blocks into existing Context Panel use cases is done upstack.

Blocks Added:
- `ActionBlock`: Renders a horizontal array of identical buttons based on an input label, icon and callback. Custom configuration options allow buttons to be hidden or disabled upon certain conditions, or to have two-step confirmation.
- `ContentBlock`: Renders any given ReactNode content. An optional prop allows the block to be collapsed.
- `ListBlock`: Renders a list of key:value pair items. Items can be with or without markers. Items without a value will not render in the list 
- `TextBlock`: Renders text. Text can be copyable or collapsible via optional props

Additionally, the `Attribute` component was added, which simply renders a given label and associated value. Value can be a string or a link.

This PR also adds unit tests to the new Blocks.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->
https://github.com/Shopify/oasis-frontend/issues/401

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Improvement
- [x] Cleanup/Refactor

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
This PR adds new UI blocks but does not implement them in the interface. If you want to test them and see what they look like visual:
- Add a test component in the code and see how it renders, or
- Go upstack to the cleanup PRs and see them in action

(wouldn't it be neat to have a UI playground?)

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->