Cleanup PipelineDetails (#1468)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

Move PipelineDetails onto UI Primitives and ContextPanel Blocks for closer visual consistency with TaskDetails.

As part of this, files were moved to a new `Context` folder.

Additionally, a new `PipelineIO` component was added, which is very similar to the existing `ArtifactsList` but slightly different in how it handles data (and uses new Block components). As such, for this PR only the `ArtifactsList` and `PipelineIO` components live side-by-side in the directory. One is consumed by PipelineDetails, the other by RunDetails. `ArtifactsList` is then removed completely when RunDetails is migrated and cleaned up in the next PR. Most of the visual aspects of `ArtifactsList` are included in `PipelineIO`.

**A note on Duplicate Efforts:**
There was no intent to create a duplicate artifact/IO list. `PipelineIO` was created before `ArtifactsList` came into being and when I compared the two to see the similarities I decided to adopt the UI of `ArtifactsList` into the functionality/structure of `PipelineIO`.

This PR also fixes some text overflow bugs in the artifact list.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

https://github.com/Shopify/oasis-frontend/issues/401

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Cleanup/Refactor

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

| Before | After |
| --- | --- |
| ![image.png](https://app.graphite.com/user-attachments/assets/28552f71-fd75-4e99-863f-9e52255774c8.png) | ![image.png](https://app.graphite.com/user-attachments/assets/e69c1a6e-f956-4f8d-97a2-2f2219d42456.png) |

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->

No change to app functionality. UI update only. Confirm that the interface works and shows info as expected.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->